+++ tests/Unit/EffectiveComponentCatalogTest.php
@@ -134,7 +134,7 @@ final class EffectiveComponentCatalogTest extends TestCase
         self::assertSame(CanonicalJson::encodePretty($first), CanonicalJson::encodePretty($second));
         self::assertSame('docara.effective_component_catalog.v1', $first['schema']);
         self::assertSame(1, $first['version']);
-        self::assertSame('ui-d81ccde2bdd5-smart-a916bbadf3aa', $first['framework_pair']);
+        self::assertSame('ui-bc8dfd7e4bdb-smart-81741eac168d', $first['framework_pair']);
         self::assertSame('4b055d09926fec4c32f2ae43b2e7e0a6f64d7663', $first['provider_revision']);
         self::assertSame([
             'catalog_is_canonical_framework_registry' => false,